+++ search/src/negamax/tt.rs
@@ -12,14 +12,14 @@ pub enum Bound {
 #[derive(Clone, Copy)]
 pub struct TTEntry {
     pub plies: u8,
-    pub value: i32,
+    pub value: i16,
     pub best_move: Option<ChessMove>,
     pub bound: Bound,
 }
 
 impl TTEntry {
     #[inline(always)]
-    pub fn new(plies: u8, value: i32, bound: Bound, best_move: Option<ChessMove>) -> Self {
+    pub fn new(plies: u8, value: i16, bound: Bound, best_move: Option<ChessMove>) -> Self {
         Self {
             plies,
             value,
